# week7/backend/src/main/java/org/openapitools/api/ProductsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Product;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-12T16:10:00.483296900+07:00[Asia/Saigon]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "products", description = "the products API")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_PRODUCTS_GET = "/products";
    /**
     * GET /products : Get all products
     *
     * @return List of products (status code 200)
     */
    @Operation(
        operationId = "productsGet",
        summary = "Get all products",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of products", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Product.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_PRODUCTS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Product>> productsGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"price\" : 1000, \"name\" : \"Laptop\", \"description\" : \"Gaming laptop\", \"id\" : \"p123\" }, { \"price\" : 1000, \"name\" : \"Laptop\", \"description\" : \"Gaming laptop\", \"id\" : \"p123\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_PRODUCTS_ID_DELETE = "/products/{id}";
    /**
     * DELETE /products/{id} : Delete product
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     */
    @Operation(
        operationId = "productsIdDelete",
        summary = "Delete product",
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ProductsApi.PATH_PRODUCTS_ID_DELETE
    )
    default ResponseEntity<Void> productsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_PRODUCTS_ID_GET = "/products/{id}";
    /**
     * GET /products/{id} : Get product by ID
     *
     * @param id  (required)
     * @return Product detail (status code 200)
     */
    @Operation(
        operationId = "productsIdGet",
        summary = "Get product by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Product detail", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_PRODUCTS_ID_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<Product> productsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : 1000, \"name\" : \"Laptop\", \"description\" : \"Gaming laptop\", \"id\" : \"p123\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_PRODUCTS_ID_PUT = "/products/{id}";
    /**
     * PUT /products/{id} : Update product
     *
     * @param id  (required)
     * @param product  (required)
     * @return Updated (status code 200)
     */
    @Operation(
        operationId = "productsIdPut",
        summary = "Update product",
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ProductsApi.PATH_PRODUCTS_ID_PUT,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> productsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_PRODUCTS_POST = "/products";
    /**
     * POST /products : Create product
     *
     * @param product  (required)
     * @return Product created (status code 201)
     */
    @Operation(
        operationId = "productsPost",
        summary = "Create product",
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ProductsApi.PATH_PRODUCTS_POST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> productsPost(
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}