# week7/backend/src/main/java/org/openapitools/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.User;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-12T16:10:00.483296900+07:00[Asia/Saigon]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_USERS_GET = "/users";
    /**
     * GET /users : Get all users
     *
     * @return List of users (status code 200)
     */
    @Operation(
        operationId = "usersGet",
        summary = "Get all users",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of users", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = User.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_USERS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<User>> usersGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"Khoe\", \"id\" : \"u123\", \"email\" : \"khoe@gmail.com\", \"age\" : 21 }, { \"name\" : \"Khoe\", \"id\" : \"u123\", \"email\" : \"khoe@gmail.com\", \"age\" : 21 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_ID_DELETE = "/users/{id}";
    /**
     * DELETE /users/{id} : Delete user
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     */
    @Operation(
        operationId = "usersIdDelete",
        summary = "Delete user",
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersApi.PATH_USERS_ID_DELETE
    )
    default ResponseEntity<Void> usersIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_ID_GET = "/users/{id}";
    /**
     * GET /users/{id} : Get user by ID
     *
     * @param id  (required)
     * @return User detail (status code 200)
     */
    @Operation(
        operationId = "usersIdGet",
        summary = "Get user by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "User detail", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_USERS_ID_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<User> usersIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"Khoe\", \"id\" : \"u123\", \"email\" : \"khoe@gmail.com\", \"age\" : 21 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_ID_PUT = "/users/{id}";
    /**
     * PUT /users/{id} : Update user
     *
     * @param id  (required)
     * @param user  (required)
     * @return Updated (status code 200)
     */
    @Operation(
        operationId = "usersIdPut",
        summary = "Update user",
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = UsersApi.PATH_USERS_ID_PUT,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> usersIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "User", description = "", required = true) @Valid @RequestBody User user
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_POST = "/users";
    /**
     * POST /users : Create user
     *
     * @param user  (required)
     * @return User created (status code 201)
     */
    @Operation(
        operationId = "usersPost",
        summary = "Create user",
        responses = {
            @ApiResponse(responseCode = "201", description = "User created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_USERS_POST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> usersPost(
        @Parameter(name = "User", description = "", required = true) @Valid @RequestBody User user
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}